+++ R/prioritise_targets_network.R
@@ -39,9 +39,7 @@
 #'
 #' @export
 #' @examples
-#' top_targets <- MSTExplorer::example_targets$top_targets[1:10]
-#' top_targets[,estimate:=fold_change]
-#' top_targets <- map_celltype(top_targets)
+#' top_targets <- example_targets$top_targets[seq(10)]
 #' vn <- prioritise_targets_network(top_targets = top_targets)
 prioritise_targets_network <- function(top_targets,
                                        vertex_vars = c("disease_name",