+++ tests/testthat/test-prioritise_targets_network.R
@@ -1,7 +1,6 @@
 test_that("prioritise_targets_network works", {
 
   top_targets <- example_targets$top_targets
-  top_targets[,estimate:=fold_change]
   top_targets <- map_celltype(top_targets)
   vn <- prioritise_targets_network(top_targets)
   testthat::expect_true(methods::is(vn$plot,"visNetwork"))